Moved code related to preferences in pref.py
